Restore upstream apt sources when rpc_release changes
Now that artifact builds and AIO PR tests are using a nodepool
image which has the RPC-O artifact repo built-in, the PR which
changes rpc_release is unable to get access to a complete set
of apt sources. This patch ensures that in the AIO bootstrap,
the sources are returned to the original upstream set if there
are no apt sources available for the given rpc_release.

JIRA: RO-4357